feat: lazy-load Editor component with Suspense boundary
